+++ pkg/manager/impl/validators/artifact_validator.go
@@ -74,7 +74,9 @@ func ValidateListArtifactRequest(request datacatalog.ListArtifactsRequest) error
 		return err
 	}
 
-	// TODO: validate PaginationOptions
+	if err := ValidatePaginationOptions(request.Pagination); err != nil {
+		return err
+	}
 
 	return nil
 }
